requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
PyMySQL==1.1.1
cryptography>=41.0.0
passlib[bcrypt]==1.7.4
bcrypt==4.3.0
Flask-Migrate==4.1.0
Flask-Login==0.6.3
Flask-Compress==1.14
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.11.0
gtts>=2.5.0
pyttsx3>=2.90